Patch para Product

+++ src/main/java/com/galonovoa/mercado/model/Product.java
@@ -3,6 +3,7 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import java.math.BigDecimal;
 
 @Entity
 public class Product {
@@ -11,12 +12,12 @@ public class Product {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String name;
-    private double price;
+    private BigDecimal price;
     private String description;
     private String image;
 
     public Product() {}
-    public Product(String name, double price, String description, String image) {
+    public Product(String name, BigDecimal price, String description, String image) {
         this.name = name;
         this.price = price;
         this.description = description;
@@ -47,11 +48,11 @@ public class Product {
         this.description = description;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
